Refactor : "모임 부분 Slice -> Page"
