+++ src/main/kotlin/org/spartaa3/movietogather/domain/meetings/dto/meetingsRequest/UpdateMeetingsRequest.kt
@@ -6,8 +6,6 @@ data class UpdateMeetingsRequest(
 
     var meetingName: String,
 
-    var movieName: String,
-
     var startTime: LocalDateTime = LocalDateTime.now(),
 
     var endTime: LocalDateTime = LocalDateTime.now(),
