+++ src/main/kotlin/org/spartaa3/movietogather/domain/meetings/repository/MeetingQueryRepository.kt
@@ -3,8 +3,8 @@ package org.spartaa3.movietogather.domain.meetings.repository
 import org.spartaa3.movietogather.domain.meetings.entity.MeetingSearchCondition
 import org.spartaa3.movietogather.domain.meetings.entity.Meetings
 import org.spartaa3.movietogather.domain.meetings.service.Type
+import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
-import org.springframework.data.domain.Slice
 
 interface MeetingQueryRepository {
     fun searchMeeting(
@@ -12,5 +12,5 @@ interface MeetingQueryRepository {
         condition: MeetingSearchCondition,
         keyword: String?,
         pageable: Pageable
-    ): Slice<Meetings>
+    ): Page<Meetings>
 }
