+++ src/main/kotlin/org/spartaa3/movietogather/domain/meetings/repository/MeetingQueryRepositoryImpl.kt
@@ -6,9 +6,9 @@ import org.spartaa3.movietogather.domain.meetings.entity.Meetings
 import org.spartaa3.movietogather.domain.meetings.entity.QMeetings
 import org.spartaa3.movietogather.domain.meetings.service.Type
 import org.spartaa3.movietogather.infra.QueryDslSupport
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.PageImpl
 import org.springframework.data.domain.Pageable
-import org.springframework.data.domain.Slice
-import org.springframework.data.domain.SliceImpl
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -19,8 +19,10 @@ class MeetingQueryRepositoryImpl : MeetingQueryRepository, QueryDslSupport() {
         condition: MeetingSearchCondition,
         keyword: String?,
         pageable: Pageable
-    ): Slice<Meetings> {
-        val pageSize = pageable.pageSize
+    ): Page<Meetings> {
+        val totalCount = queryFactory
+            .selectFrom(meetings)
+            .where(allCond(type, condition, keyword)).fetch().size.toLong()
         val contents = queryFactory
             .selectFrom(meetings)
             .where(allCond(type, condition, keyword))
@@ -29,13 +31,8 @@ class MeetingQueryRepositoryImpl : MeetingQueryRepository, QueryDslSupport() {
             .orderBy(meetings.startTime.asc())
             .fetch()
 
-        var hasNext = false
-        if (contents.size > pageSize) {
-            contents.removeAt(pageSize)
-            hasNext = true
-        }
 
-        return SliceImpl(contents, pageable, hasNext)
+        return PageImpl(contents, pageable, totalCount)
     }
 
     private fun allCond(type: Type, condition: MeetingSearchCondition, keyword: String?): BooleanExpression? {
