+++ src/main/kotlin/org/spartaa3/movietogather/global/cookie/CookieUtils.kt
@@ -50,7 +50,10 @@ object CookieUtils {
     }
 
     fun <T> deserialize(cookie: Cookie, cls: Class<T>): T {
-        return cls.cast(SerializationUtils.deserialize(
-            Base64.getUrlDecoder().decode(cookie.value)))
+        return cls.cast(
+            SerializationUtils.deserialize(
+                Base64.getUrlDecoder().decode(cookie.value)
+            )
+        )
     }
 }
