+++ src/main/kotlin/org/spartaa3/movietogather/infra/security/jwt/JwtAuthenticationFilter.kt
@@ -3,7 +3,6 @@ package org.spartaa3.movietogather.infra.security.jwt
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import org.spartaa3.movietogather.global.cookie.CookieUtils
 import org.springframework.http.HttpHeaders
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
@@ -13,10 +12,11 @@ import org.springframework.web.filter.OncePerRequestFilter
 @Component
 class JwtAuthenticationFilter(
     private val jwtPlugin: JwtPlugin
-): OncePerRequestFilter() {
-    companion object{
+) : OncePerRequestFilter() {
+    companion object {
         private val BEARER_PATTERN = Regex("^Bearer (.+?)$")
     }
+
     override fun doFilterInternal(
         request: HttpServletRequest,
         response: HttpServletResponse,
@@ -45,6 +45,7 @@ class JwtAuthenticationFilter(
         }
         filterChain.doFilter(request, response)
     }
+
     private fun HttpServletRequest.getBearerToken(): String? {
         val headerValue = this.getHeader(HttpHeaders.AUTHORIZATION) ?: return null
         return BEARER_PATTERN.find(headerValue)?.groupValues?.get(1)
