+++ src/test/kotlin/org/spartaa3/movietogather/MovieToGatherApplicationTests.kt
@@ -1,4 +1,4 @@
-package org.spartaa3.movietogather.domain.review
+package org.spartaa3.movietogather
 
 import org.springframework.boot.test.context.SpringBootTest
